Some metadata improvements
* Fix README.rst to render on PyPI (figure instead of image)
* Make link to docs HTTPS
* Make changelog reverse order and in restructured text, included on PyPI page for easy access
* Automate putting README on to PyPI in setup.py
* Fix `setup.py` to include project URL pointing to Github
* Remove unnecessary brackets around email

`python setup.py check -s --restructuredtext --metadata` tests your setup.py and the validity of the reStructuredText in the `long_description` field, it used to fail but after this change passes.

If you run `python setup.py register` after this is merged it will instantly update the PyPI page without needing to wait for the next release.