Changing service time.
The previous service time was calculated by using a sample number,
that has been changed to be calculated as ending time minus the
starting time.

In the nodes parameters we are no longer passing the service time,
but the end date service to calculate the service time.

The test had to be altered as well.