Interurpted individuals get False instead of None
False service_time and service_end_date at interrutions.
This is in order to be able to compare False with other numbers. Also bring in line with customers who have yet to receive a service_end_date. Logic does not fail as a False service_end_date and a service_end_date before the current time are treated the same.