Attempt to implement context manager (do not merge)
This isn't working: see failing test.

Essentially I can't figure out how to delete an item using the pointer
in the list:

```python
>>> stuff = [1, 2, 3, 4]
>>> vince = stuff[2]
>>> del stuff[:]
>>> stuff, vince
([], 3)

```

I've read quite a few things that essentially say: "if you're trying to
do this, try to do something else". For example
https://stackoverflow.com/questions/1316767/how-can-i-explicitly-free-memory-in-python
(although the advice of using generators doesn't apply here).

I believe this is inherent to how ciw is built and that there's
potentially quite a few cyclic connections.

If we can't figure out how to delete an element inside a list I'd
suggest going down the line of not having classes for customers at all
(named tuples as you suggested Geraint) and writing out to file.

I think moving to named tuples shouldn't be too hard (it's essentially
what the classes are for here) and then subsequently writing out to file
would be a bit of work.