add a fast path to all_individuals for simulations with one priority class
`all_individuals` creates a new list on each invocation, which is a problem because it is so frequently called: flatten_list() alone takes up %% of execution time. I first added a `len(list_of_lists) == 1` check there, but it still showed up in the trace, due to the overhead of (I presume) the function call, so I moved the check inside `all_individuals`.