Update tracker parameter in simulation.py
Default argument was a mutable type. Generally the advice I've read is to only use immutable types as defaults. Although I was not running into this issue, it can in principle lead to unusual behavior when a default argument is mutable.

In this case I have changed the default argument of `tracker` to `None`, but when this happens the same tracker will be assigned as before.