customer_class is hashable
Because `customer_class` merely needs to be hashable and ordered, and there is no currently-supported typing for being ordered in Python, setting it to hashable seems like the closest typing to use for the current state. 

I have actually been using dataclasses with `@dataclass(order=True, frozen=True)` so that I can keep track of a variety of types of data that distinguish classes of patients based on their attributes. But I wouldn't want to make defining dataclasses a requirement.